+++ lib/exceptions.rb
@@ -1,0 +1,3 @@
+module Fb
+  class EmergencyStop < StandardError; end
+end
